refactor(import): pull out nightbot types and fns
